feat(it01h01): add geocoding with leaflet-control-geocoder
