+++ my-app/src/app/components/map/map.component.html
@@ -5,6 +5,6 @@
 </div>
 <div class="contenedor-ruta">
   <div>
-    <button (click)="calcularRuta()" class="boton-grande">Calcular Ruta</button>
+    <button (click)="calculateRoute()" class="boton-grande">Calcular Ruta</button>
   </div>
 </div>
